# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SymptomTag = (typeof SymptomTag)[keyof typeof SymptomTag];

export const SymptomTag = {
  cravings: "cravings",
  bloating: "bloating",
  fatigue: "fatigue",
  anxiety: "anxiety",
  sad: "sad",
  happy: "happy",
  energetic: "energetic",
  cramps: "cramps",
  headache: "headache",
  acne: "acne",
  insomnia: "insomnia",
  soreness: "soreness",
  horny: "horny",
} as const;

export type HabitTag = (typeof HabitTag)[keyof typeof HabitTag];

export const HabitTag = {
  water: "water",
  walk: "walk",
  stretch: "stretch",
  meditate: "meditate",
  protein: "protein",
  sleep: "sleep",
} as const;

export type CyclePhase = (typeof CyclePhase)[keyof typeof CyclePhase];

export const CyclePhase = {
  menstrual: "menstrual",
  follicular: "follicular",
  ovulatory: "ovulatory",
  luteal: "luteal",
} as const;

export type ConcernKey = (typeof ConcernKey)[keyof typeof ConcernKey];

export const ConcernKey = {
  acne: "acne",
  "hair-loss": "hair-loss",
  hirsutism: "hirsutism",
  "irregular-cycle": "irregular-cycle",
  "weight-belly": "weight-belly",
  fatigue: "fatigue",
  cravings: "cravings",
  insomnia: "insomnia",
  anxiety: "anxiety",
  "dark-patches": "dark-patches",
} as const;

export interface AnalyzeVentContext {
  phase: CyclePhase;
  /** @minimum 1 */
  dayOfCycle: number;
  /** @minimum 14 */
  cycleLength: number;
  diet: string;
  homeCountry: string;
  travelling: boolean;
  travelCountry?: string | null;
  /**
   * @minimum 1
   * @maximum 5
   */
  energy: number;
  trackedConcerns: ConcernKey[];
  /** @maxItems 5 */
  recentVentTexts: string[];
}

export interface AnalyzeVentRequest {
  /**
   * @minLength 2
   * @maxLength 2000
   */
  text: string;
  context: AnalyzeVentContext;
}

export type VentAnalysisFollowUp = {
  concern: ConcernKey;
  label: string;
} | null;

export interface VentAnalysis {
  symptoms: SymptomTag[];
  habits: HabitTag[];
  headline: string;
  explanation: string;
  followUp?: VentAnalysisFollowUp;
}

export type VibeMood = (typeof VibeMood)[keyof typeof VibeMood];

export const VibeMood = {
  low: "low",
  anxious: "anxious",
  vibrant: "vibrant",
  steady: "steady",
} as const;

export type VibeTaskKind = (typeof VibeTaskKind)[keyof typeof VibeTaskKind];

export const VibeTaskKind = {
  rest: "rest",
  hydration: "hydration",
  movement: "movement",
  food: "food",
  social: "social",
  mindset: "mindset",
  supplement: "supplement",
} as const;

export interface InterpretVibeRequest {
  /**
   * @minLength 1
   * @maxLength 280
   */
  sentence: string;
  context: AnalyzeVentContext;
}

export interface VibeInjectedTask {
  title: string;
  detail: string;
  why: string;
  kind: VibeTaskKind;
}

export type VibePaletteIntensity =
  (typeof VibePaletteIntensity)[keyof typeof VibePaletteIntensity];

export const VibePaletteIntensity = {
  soft: "soft",
  normal: "normal",
  punchy: "punchy",
} as const;

/**
 * Full server-driven directive for reshaping the home screen. The
server controls every layout decision so the client renders this
verbatim — it does not re-derive layout from mood alone.

 */
export interface VibeDirective {
  mood: VibeMood;
  /**
   * How strongly to apply (1 mild, 3 strong).
   * @minimum 1
   * @maximum 3
   */
  intensity: number;
  /** Short warm sentence shown above the home (max ~18 words). */
  headline: string;
  /** One short sentence explaining why we changed the home, plain language. */
  explanation: string;
  /** Tiny label for the "Tuned for …" pill (max ~6 words). */
  pillLabel: string;
  /** Section header for the mission section. */
  missionTitle: string;
  /** Section sub-copy for the mission section. */
  missionSub: string;
  /** Optional sub-line for the streak chip (e.g. for low mood). */
  streakNote: string | null;
  injectedTask: VibeInjectedTask | null;
  /**
   * How aggressively to strip away non-essentials. 0 = full home, 1 = hide bonus
sections, 2 = mission only + minimum, 3 = rest mode.

   * @minimum 0
   * @maximum 3
   */
  simplifyLevel: number;
  /** Task kinds to drop entirely from the home today. */
  hideTaskKinds: VibeTaskKind[];
  /** If true, replace any remaining movement task with a restorative stretch. */
  swapMovementToRest: boolean;
  /** If true, render the Lean-in section above the mission section. */
  promoteLean: boolean;
  paletteIntensity: VibePaletteIntensity;
  /** When the directive should auto-clear and the home returns to baseline. */
  expiresAt: string;
}

export type VentAnalysisErrorCode =
  (typeof VentAnalysisErrorCode)[keyof typeof VentAnalysisErrorCode];

export const VentAnalysisErrorCode = {
  bad_request: "bad_request",
  llm_unavailable: "llm_unavailable",
  llm_timeout: "llm_timeout",
  internal: "internal",
} as const;

export interface VentAnalysisError {
  code: VentAnalysisErrorCode;
  message: string;
}

export type AuthProvider = (typeof AuthProvider)[keyof typeof AuthProvider];

export const AuthProvider = {
  apple: "apple",
  google: "google",
  dev: "dev",
} as const;

export interface UserProfile {
  id: number;
  provider: AuthProvider;
  /** @nullable */
  email: string | null;
  /** @nullable */
  name: string | null;
  /** @nullable */
  diet: string | null;
  /** @nullable */
  homeCountry: string | null;
  /** @nullable */
  energy: number | null;
  /** @nullable */
  lastPeriodIso: string | null;
  /** @nullable */
  cycleLength: number | null;
  /** @nullable */
  onboardedAt: string | null;
}

export interface AuthSession {
  token: string;
  user: UserProfile;
}

export interface AppleSignInInput {
  /** @minLength 16 */
  identityToken: string;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  email?: string | null;
}

export interface GoogleSignInInput {
  /** @minLength 16 */
  idToken: string;
}

export interface DevSignInInput {
  /**
   * @minLength 1
   * @maxLength 64
   */
  handle: string;
}

export interface ProfileUpdateInput {
  /**
   * @minLength 1
   * @maxLength 80
   */
  name?: string;
  /**
   * @minLength 1
   * @maxLength 40
   */
  diet?: string;
  /**
   * @minLength 1
   * @maxLength 8
   */
  homeCountry?: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  energy?: number;
  /**
   * @minLength 8
   * @maxLength 40
   */
  lastPeriodIso?: string;
  /**
   * @minimum 14
   * @maximum 60
   */
  cycleLength?: number;
  onboarded?: boolean;
}

export type AuthErrorCode = (typeof AuthErrorCode)[keyof typeof AuthErrorCode];

export const AuthErrorCode = {
  bad_request: "bad_request",
  unauthorized: "unauthorized",
  provider_unconfigured: "provider_unconfigured",
  not_found: "not_found",
  internal: "internal",
} as const;

export interface AuthError {
  code: AuthErrorCode;
  message: string;
}