# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: vents
    description: Vent journal LLM analysis
  - name: vibe
    description: One-sentence "how do I feel" → home-screen directive
  - name: auth
    description: Apple/Google sign-in and user profile
paths:
  /auth/apple:
    post:
      operationId: signInWithApple
      tags: [auth]
      summary: Exchange a verified Apple identity token for a Lumen session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppleSignInInput"
      responses:
        "200":
          description: Session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "401":
          description: Identity token rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "503":
          description: Auth provider misconfigured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
  /auth/google:
    post:
      operationId: signInWithGoogle
      tags: [auth]
      summary: Exchange a verified Google ID token for a Lumen session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleSignInInput"
      responses:
        "200":
          description: Session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "401":
          description: Identity token rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "503":
          description: Auth provider misconfigured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
  /auth/dev:
    post:
      operationId: signInDev
      tags: [auth]
      summary: |
        Development-only sign-in for the Replit web preview where native
        Apple/Google flows are unavailable. Disabled in production.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DevSignInInput"
      responses:
        "200":
          description: Session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "404":
          description: Not enabled in this environment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      security:
        - bearerAuth: []
      summary: Return the current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
  /auth/signout:
    post:
      operationId: signOut
      tags: [auth]
      security:
        - bearerAuth: []
      summary: Invalidate the current session for this user
      responses:
        "204":
          description: Signed out
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      security:
        - bearerAuth: []
      summary: Persist onboarding interview answers onto the user record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdateInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /vents/analyze:
    post:
      operationId: analyzeVent
      tags: [vents]
      summary: Analyze a vent transcript
      description: |
        Runs LLM analysis on a vent transcript using the user's PCOS context
        (cycle phase, diet, energy, tracked concerns, recent vents) and returns
        structured detections plus a phase-aware coaching response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeVentRequest"
      responses:
        "200":
          description: Structured analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VentAnalysis"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VentAnalysisError"
        "503":
          description: LLM upstream unavailable; client should fall back
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VentAnalysisError"
        "504":
          description: LLM upstream timed out; client should fall back
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VentAnalysisError"
  /vibe/interpret:
    post:
      operationId: interpretVibe
      tags: [vibe]
      summary: Interpret a one-sentence mood and return a home-screen directive
      description: |
        Takes one short sentence about how the user is feeling right now plus
        her PCOS context and returns a structured "vibe directive" the home
        screen can use to reshape itself: mood category, intensity, layout
        copy, an optional micro-task to prepend, and an expiry timestamp.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterpretVibeRequest"
      responses:
        "200":
          description: Vibe directive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VibeDirective"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VentAnalysisError"
        "503":
          description: LLM upstream unavailable; client should fall back
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VentAnalysisError"
        "504":
          description: LLM upstream timed out; client should fall back
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VentAnalysisError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SymptomTag:
      type: string
      enum:
        - cravings
        - bloating
        - fatigue
        - anxiety
        - sad
        - happy
        - energetic
        - cramps
        - headache
        - acne
        - insomnia
        - soreness
        - horny
    HabitTag:
      type: string
      enum:
        - water
        - walk
        - stretch
        - meditate
        - protein
        - sleep
    CyclePhase:
      type: string
      enum: [menstrual, follicular, ovulatory, luteal]
    ConcernKey:
      type: string
      enum:
        - acne
        - hair-loss
        - hirsutism
        - irregular-cycle
        - weight-belly
        - fatigue
        - cravings
        - insomnia
        - anxiety
        - dark-patches
    AnalyzeVentContext:
      type: object
      properties:
        phase:
          $ref: "#/components/schemas/CyclePhase"
        dayOfCycle:
          type: integer
          minimum: 1
        cycleLength:
          type: integer
          minimum: 14
        diet:
          type: string
        homeCountry:
          type: string
        travelling:
          type: boolean
        travelCountry:
          type: string
          nullable: true
        energy:
          type: integer
          minimum: 1
          maximum: 5
        trackedConcerns:
          type: array
          items:
            $ref: "#/components/schemas/ConcernKey"
        recentVentTexts:
          type: array
          items:
            type: string
          maxItems: 5
      required:
        - phase
        - dayOfCycle
        - cycleLength
        - diet
        - homeCountry
        - travelling
        - energy
        - trackedConcerns
        - recentVentTexts
    AnalyzeVentRequest:
      type: object
      properties:
        text:
          type: string
          minLength: 2
          maxLength: 2000
        context:
          $ref: "#/components/schemas/AnalyzeVentContext"
      required:
        - text
        - context
    VentAnalysis:
      type: object
      properties:
        symptoms:
          type: array
          items:
            $ref: "#/components/schemas/SymptomTag"
        habits:
          type: array
          items:
            $ref: "#/components/schemas/HabitTag"
        headline:
          type: string
        explanation:
          type: string
        followUp:
          type: object
          nullable: true
          properties:
            concern:
              $ref: "#/components/schemas/ConcernKey"
            label:
              type: string
          required:
            - concern
            - label
      required:
        - symptoms
        - habits
        - headline
        - explanation
    VibeMood:
      type: string
      enum: [low, anxious, vibrant, steady]
    VibeTaskKind:
      type: string
      enum: [rest, hydration, movement, food, social, mindset, supplement]
    InterpretVibeRequest:
      type: object
      properties:
        sentence:
          type: string
          minLength: 1
          maxLength: 280
        context:
          $ref: "#/components/schemas/AnalyzeVentContext"
      required:
        - sentence
        - context
    VibeInjectedTask:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        why:
          type: string
        kind:
          $ref: "#/components/schemas/VibeTaskKind"
      required:
        - title
        - detail
        - why
        - kind
    VibePaletteIntensity:
      type: string
      enum: [soft, normal, punchy]
    VibeDirective:
      type: object
      description: |
        Full server-driven directive for reshaping the home screen. The
        server controls every layout decision so the client renders this
        verbatim — it does not re-derive layout from mood alone.
      properties:
        mood:
          $ref: "#/components/schemas/VibeMood"
        intensity:
          type: integer
          minimum: 1
          maximum: 3
          description: How strongly to apply (1 mild, 3 strong).
        headline:
          type: string
          description: Short warm sentence shown above the home (max ~18 words).
        explanation:
          type: string
          description: One short sentence explaining why we changed the home, plain language.
        pillLabel:
          type: string
          description: Tiny label for the "Tuned for …" pill (max ~6 words).
        missionTitle:
          type: string
          description: Section header for the mission section.
        missionSub:
          type: string
          description: Section sub-copy for the mission section.
        streakNote:
          type: string
          nullable: true
          description: Optional sub-line for the streak chip (e.g. for low mood).
        injectedTask:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/VibeInjectedTask"
        simplifyLevel:
          type: integer
          minimum: 0
          maximum: 3
          description: |
            How aggressively to strip away non-essentials. 0 = full home, 1 = hide bonus
            sections, 2 = mission only + minimum, 3 = rest mode.
        hideTaskKinds:
          type: array
          items:
            $ref: "#/components/schemas/VibeTaskKind"
          description: Task kinds to drop entirely from the home today.
        swapMovementToRest:
          type: boolean
          description: If true, replace any remaining movement task with a restorative stretch.
        promoteLean:
          type: boolean
          description: If true, render the Lean-in section above the mission section.
        paletteIntensity:
          $ref: "#/components/schemas/VibePaletteIntensity"
        expiresAt:
          type: string
          format: date-time
          description: When the directive should auto-clear and the home returns to baseline.
      required:
        - mood
        - intensity
        - headline
        - explanation
        - pillLabel
        - missionTitle
        - missionSub
        - streakNote
        - injectedTask
        - simplifyLevel
        - hideTaskKinds
        - swapMovementToRest
        - promoteLean
        - paletteIntensity
        - expiresAt
    VentAnalysisError:
      type: object
      properties:
        code:
          type: string
          enum: [bad_request, llm_unavailable, llm_timeout, internal]
        message:
          type: string
      required:
        - code
        - message
    AuthProvider:
      type: string
      enum: [apple, google, dev]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        provider:
          $ref: "#/components/schemas/AuthProvider"
        email:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        diet:
          type: ["string", "null"]
        homeCountry:
          type: ["string", "null"]
        energy:
          type: ["integer", "null"]
        lastPeriodIso:
          type: ["string", "null"]
        cycleLength:
          type: ["integer", "null"]
        onboardedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - provider
        - email
        - name
        - diet
        - homeCountry
        - energy
        - lastPeriodIso
        - cycleLength
        - onboardedAt
    AuthSession:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    AppleSignInInput:
      type: object
      properties:
        identityToken:
          type: string
          minLength: 16
        fullName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
      required:
        - identityToken
    GoogleSignInInput:
      type: object
      properties:
        idToken:
          type: string
          minLength: 16
      required:
        - idToken
    DevSignInInput:
      type: object
      properties:
        handle:
          type: string
          minLength: 1
          maxLength: 64
      required:
        - handle
    ProfileUpdateInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        diet:
          type: string
          minLength: 1
          maxLength: 40
        homeCountry:
          type: string
          minLength: 1
          maxLength: 8
        energy:
          type: integer
          minimum: 1
          maximum: 10
        lastPeriodIso:
          type: string
          minLength: 8
          maxLength: 40
        cycleLength:
          type: integer
          minimum: 14
          maximum: 60
        onboarded:
          type: boolean
    AuthError:
      type: object
      properties:
        code:
          type: string
          enum:
            - bad_request
            - unauthorized
            - provider_unconfigured
            - not_found
            - internal
        message:
          type: string
      required:
        - code
        - message
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT