# lib/api-zod/src/generated/types/appleSignInInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AppleSignInInput {
  /** @minLength 16 */
  identityToken: string;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  email?: string | null;
}