# lib/api-zod/src/generated/types/authError.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthErrorCode } from "./authErrorCode";

export interface AuthError {
  code: AuthErrorCode;
  message: string;
}