# lib/api-zod/src/generated/types/authProvider.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type AuthProvider = (typeof AuthProvider)[keyof typeof AuthProvider];

export const AuthProvider = {
  apple: "apple",
  google: "google",
  dev: "dev",
} as const;