# lib/api-zod/src/generated/types/devSignInInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DevSignInInput {
  /**
   * @minLength 1
   * @maxLength 64
   */
  handle: string;
}