# lib/api-zod/src/generated/types/profileUpdateInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdateInput {
  /**
   * @minLength 1
   * @maxLength 80
   */
  name?: string;
  /**
   * @minLength 1
   * @maxLength 40
   */
  diet?: string;
  /**
   * @minLength 1
   * @maxLength 8
   */
  homeCountry?: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  energy?: number;
  /**
   * @minLength 8
   * @maxLength 40
   */
  lastPeriodIso?: string;
  /**
   * @minimum 14
   * @maximum 60
   */
  cycleLength?: number;
  onboarded?: boolean;
}