# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthProvider } from "./authProvider";

export interface UserProfile {
  id: number;
  provider: AuthProvider;
  /** @nullable */
  email: string | null;
  /** @nullable */
  name: string | null;
  /** @nullable */
  diet: string | null;
  /** @nullable */
  homeCountry: string | null;
  /** @nullable */
  energy: number | null;
  /** @nullable */
  lastPeriodIso: string | null;
  /** @nullable */
  cycleLength: number | null;
  /** @nullable */
  onboardedAt: Date | null;
}